+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version 1.7.0 (not yet released)
 
-
+* Changed Age field so aggeragation is working correctly.
+* Changed Field Specification to allow more advanced mysql functions
 
 # Version 1.6.0